Added jesd_status_simple Example
Added a simple example application which leverages
the jesd_common tools for reading device and lane
information, and simply prints the results to
the terminal.  This application intends to just
show how to access the data via the application,
as a base for those wanting to monitor JESD status
within their own custom applications.

Signed-off-by: Brent Kowal <brent.kowal@analog.com>